inboxbench/requirements.txt
requests
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
resend

